rename query module
