+++ hs-abci-examples/nameservice/src/Nameservice/Modules/Nameservice/Query.hs
@@ -5,10 +5,8 @@ import           Nameservice.Modules.Nameservice.Keeper (storeKey)
 import           Nameservice.Modules.Nameservice.Types  (Name, Whois)
 import           Polysemy                               (Member, Sem)
 import           Servant.API                            ((:>))
-import qualified Tendermint.SDK.Router                  as R
+import qualified Tendermint.SDK.Query                   as Q
 import qualified Tendermint.SDK.Store                   as Store
-import           Tendermint.SDK.StoreQueries            (QueryApi,
-                                                         storeQueryHandlers)
 
 --------------------------------------------------------------------------------
 -- | Query API
@@ -16,8 +14,8 @@ import           Tendermint.SDK.StoreQueries            (QueryApi,
 
 type NameserviceContents = '[(Name, Whois)]
 
-type Api = "nameservice" :> QueryApi NameserviceContents
+type Api = "nameservice" :> Q.QueryApi NameserviceContents
 
-server :: Member Store.RawStore r => R.RouteT Api (Sem r)
+server :: Member Store.RawStore r => Q.RouteT Api (Sem r)
 server =
-  storeQueryHandlers (Proxy :: Proxy NameserviceContents) storeKey (Proxy :: Proxy (Sem r))
+  Q.storeQueryHandlers (Proxy :: Proxy NameserviceContents) storeKey (Proxy :: Proxy (Sem r))
