+++ hs-abci-examples/nameservice/src/Nameservice/Modules/Token.hs
@@ -52,10 +52,10 @@ import           Tendermint.SDK.Codec         (HasCodec (..))
 import           Tendermint.SDK.Errors        (AppError (..), IsAppError (..))
 import           Tendermint.SDK.Events        (Event, FromEvent, ToEvent (..),
                                                emit)
-import           Tendermint.SDK.Router        (EncodeQueryResult, FromQueryData,
-                                               Queryable (..), RouteT)
+import           Tendermint.SDK.Query         (FromQueryData, Queryable (..),
+                                               RouteT)
+import           Tendermint.SDK.Query.Store   (QueryApi, storeQueryHandlers)
 import qualified Tendermint.SDK.Store         as Store
-import           Tendermint.SDK.StoreQueries  (QueryApi, storeQueryHandlers)
 import           Tendermint.SDK.Types.Address (Address, addressFromBytes,
                                                addressToBytes)
 
@@ -117,8 +117,6 @@ instance ToEvent Transfer where
   makeEventType _ = "Transfer"
 
 instance FromEvent Transfer
-
-instance EncodeQueryResult Amount
 
 --------------------------------------------------------------------------------
 -- Exceptions
