+++ hs-abci-examples/simple-storage/src/SimpleStorage/Application.hs
@@ -1,5 +1,3 @@
-{-# LANGUAGE UndecidableInstances #-}
-
 module SimpleStorage.Application
   ( AppError(..)
   , AppConfig(..)
