+++ hs-abci-sdk/src/Tendermint/SDK/Query/Delayed.hs
@@ -1,4 +1,4 @@
-module Tendermint.SDK.Router.Delayed where
+module Tendermint.SDK.Query.Delayed where
 
 import           Control.Error                        (ExceptT, runExceptT)
 import           Control.Monad.Reader                 (MonadReader, ReaderT,
@@ -9,11 +9,12 @@ import           Data.Default.Class                   (def)
 import           Data.String.Conversions              (cs)
 import qualified Network.ABCI.Types.Messages.Request  as Request
 import qualified Network.ABCI.Types.Messages.Response as Response
-import           Tendermint.SDK.Router.Types          (QueryError (..),
+import           Tendermint.SDK.Query.Types           (QueryError (..),
                                                        RouteResult (..),
                                                        RouteResultT (..))
 
 --------------------------------------------------------------------------------
+-- NOTE: most of this was vendored and repurposed from servant
 
 
 newtype DelayedM m a =
